Initial push. All 3 algorithms and command line reading
All 3 of the algorithms in the DAME paper are in dame_algorithms.py. The first one "Algorithm1" does the initial match on all covariates, and iterates through the dataframe until there is an empty set of treatment groups. Algorithm 2 calls FLAME to group. Algorithm 3 finds the new covariate set. The main starting point of the code is in DAME.py with the command line parsing. Some test cases are in *_test.py, along with two sample csv files.