bug fix, and added adaptive-weights version that finds covariate to drop based on ridge regression
So found/fixed a bug in the generate new active sets code. Will want to come back and test this a little bit more, but I think it's fine now for 1 character column names. Added the code that allows for ridge regression minimum PE to decide which covar to drop and tested it a bit.