Adds new post processing function for variance of ATE and bug fixes  (#59)
* Update .travis.yml

copied Tommy's travis file in attempt to get the windows/mac setup going

Co-Authored-By: Thomas Howell <65472983+thowell332@users.noreply.github.com>

* Update requirements.txt

updated numpy higher in attempt to get it to work on python 3.7 in linux

* Update .travis.yml

changed pip to pip3 in attempt to get mac to work

* Update .travis.yml

there's a python 3.9 apparently ugh

* Update .travis.yml

add installation of nose to hopefully get it to work on mac
also added python 3.9 haha

* Update .travis.yml

update nose installation to --user to try to get it to work for mac still

* try to get 3.65 to work on windows

adjust windows 3.6 to 3.6.5

change language to generic to try to get macos to work.

* Update .travis.yml

now running install nose without pip3

* Update .travis.yml

removed macos and windows 2.6 attempts, done trying to fix that lol.

now trying to get it configured to run all versions of scikit with all versions of python

* Update .travis.yml

put pip3 back to try to get windows to work again

* Update .travis.yml

try to define environment variables differently so it works on windows too again

* Update .travis.yml

try mixing numpy versions and scikit versions and see what happens

* Update .travis.yml

not sure why numpy not going

* Update .travis.yml

another attempt to get numpy to vary

* Update .travis.yml

just realized that scikit uses numpy, scipy and pandas, so maybe we really only need scikit as a dependency??

* trying to change requirements

hmm  maybe that doesnt work??

* make all windows run both sklearn versions

hopefully the final run

* whoops typo

no dashes I guess?

* changing pe_frac to pe_val because it doesn't need to be bounded 0->1

Thanks to Nick for pointing out this issue

* correct the settings files

* Bug fixes

Off by one in stopping criteria for DAME, so factored it out of dame_flame_helpers.
Also, initial steps to flagging the overflow issue, but this needs to be more seriously dealt with

* Move to numpy arrays in bit manip, error trap for 0 columns needing to be binarized

Moved to numpy arrays to make the size of dataframe that can be handled larger. The error on binarization was in the step to make decisiontrees able to handle categorical data...we need to check that there exist columns that need be binarized prior to binarizing them

* Rollback all changes

To match the dame-flame repo up on the almost-matching-exactly acccount

* Add epsilon PE change

Per requested by Cynthia, this changes the default stopping criteria to stop if the PE jumps by too much.

* Update FLAME_results_test.py

Debugging the possible linux test failure??

* Update dame_algorithm.py

Debugging linux test failure still

* Update FLAME_results_test.py

Debugging Linux failure in DAME return values

* Update post_processing.py

debugging MG function because that appears to cause Linux error only.....??!!?!

* Update post_processing.py

isolating linux issue still

* Update post_processing.py

Still isolating MG issue for linux

* Update post_processing.py

Was the issue the double quotes vs single quotes??? :O

* Update post_processing.py

Apparently the columns are not found in linux...??

* Update post_processing.py

Perhaps the error is caused by double brackets vs single??

* Remove debugging print statements

not needed

* Enhancement to handle larger datasets using numpy arrays

and possible bug fix in the case that someone selects decision trees but has no columns that need to be binarized

* Update .travis.yml

Trying to fix broken tests due to coveralls code coverage checking

* Update .travis.yml

Again trying to fix broken coverall tests for code coverage

* Update .travis.yml

3rd try to fix tests via pip install coveralls

* Update .travis.yml

moved the --user to the other part of the command

* Update .travis.yml

moved the coveralls to before_install whatever that does

* Update .travis.yml

aaand back to trying pip3

* Update README.md

change travis .org to .com

* Rough variance function

This is a first pass at an implementation of the variance estimator from Abadie et al. It runs, and it's not orders of magnitude incorrect in trials compared to R-FLAME but needs a little more testing, and should be used in conjunction with the Abadie et al ATE estimator too

* Whoops pushed config files and csv

deleting unnecessary files whoops

* updated correct code for var estimator

* Update post_processing.py

Removed unnecessary print statement and return value, abbreviated notation of the treatment col name

* Travis testing issue possible fixes

For some reason the travis build is failing on py.8 windows sklearn 0.23.2
And also update the readme again to correctly connect the travis build (I thought I already did this?? weird)

* Add test for new variance function

And also flip the variance ate return order and also try to make the readme display the travis badge correctly again

* Update FLAME_results_test.py

Oops silly copy paste error in the test

* Update FLAME_results_test.py

Silly rounding mistake

* Add STAR example

In the examples folder

* Delete .DS_Store

Pretty sure this is a mac settings file that shouldn't have made it up to the main repo

* Update comments in var_ATE function

the new var_ATE function had some weird comments to remove

---------

Co-authored-by: N. Gupta <3536509+guptane6@users.noreply.github.com>
Co-authored-by: Thomas Howell <65472983+thowell332@users.noreply.github.com>