Off by one (#60)
* remove db code

* fixed off by one errors

Iterations start at 0. An iteration is completed once we have successfully attempted the match for that iteration. Also fixed pe_stopping and adjusted defaults for pre_dame and early_stop_iterations to be float('inf')

* eliminated manual call to unittest

* removed flame_db tests

* update documentation

early_stop_pe_frac is now eliminated and documentation is updated as it pertains to iterations.

---------

Co-authored-by: Vittorio Orlandi <52432620+vittorioorlandi@users.noreply.github.com>

Per @vittorioorlandi

1. Fixed the definition of early stopping with respect to PE frac. Previously it was comparing to the previous iteration PE when it should be comparing to the PE at baseline.
Accordingly, the PE at the 0'th iteration is always the PE using all covariates — even if we make no exact matches.

2. Have everything starting at 0, with a completed iteration corresponding to successfully attempting matching.
Updated defaults for pre_dame and early_stop_iterations; they are now float(‘inf’). This makes sense if they are supposed to be the number of iterations before something happens; if they are infinite they never happen (ie never early stop because of iteration and never switch to dame).

3. Updated some notebooks to reflect the above.

4. Eliminated the database folder and corresponding test folder for it -- this is temporary to get it up on PyPi with all the recent updates but will be put back shortly.

5. Updates documentation to reflect the above

This takes care of #47 and #48 without breaking #32 or #33